+++ CHANGES.rst
@@ -391,6 +391,8 @@ astropy.units
   ``to`` and ``to_value`` methods can be properly used to initialize ``Quantity``
   or set ``Quantity`` instance items. [#8535]
 
+- Add support for ``<<`` to create logarithmic units. [#8290]
+
 astropy.utils
 ^^^^^^^^^^^^^